requirements.txt
flask
langchain
faiss-cpu
openai
google-generativeai
python-dotenv
flask-cors
